requirements.txt
opencv-python
mediapipe
numpy
spotipy
pyttsx3
python-dotenv
